feature: add vao

+++ src/app/shaders/vertex-shader.glsl
@@ -2,7 +2,7 @@
 
 in vec3 a_position;
 
-uniform float uPointSize;
+uniform mediump float uPointSize;
 uniform float uAngle;
 
 void main() {
